Tratamento de exceções

+++ dao/UserDao.php
@@ -16,13 +16,11 @@ class UserDao
     {
         $query_users = "SELECT id FROM users;";
         $stmt_users = $this->db_conn->prepare($query_users);
-        $stmt_users->execute();
 
-        $result = $stmt_users->fetchAll(PDO::FETCH_ASSOC);
+        if ($stmt_users->execute()) {
+            $result = $stmt_users->fetchAll(PDO::FETCH_ASSOC);
 
-        $users_list = [];
-
-        if ($result) {
+            $users_list = [];
 
             foreach ($result as $user_id) {
                 array_push($users_list, $this->getUserById((int) $user_id['id']));
@@ -31,8 +29,10 @@ class UserDao
             return $users_list;
 
         } else {
-            throw new Exception("No user found in database");
+            throw new Exception("Erro ao obter lista de usuários!");
         }
+
+
     }
 
     /**
@@ -45,32 +45,35 @@ class UserDao
         $query_user = "SELECT * FROM users WHERE id = :user_id LIMIT 1";
         $stmt_user = $this->db_conn->prepare($query_user);
         $stmt_user->bindParam(':user_id', $user_id, PDO::PARAM_INT);
-        $stmt_user->execute();
 
-        $result_user = $stmt_user->fetch(PDO::FETCH_ASSOC);
+        if ($stmt_user->execute()) {
+            $result_user = $stmt_user->fetch(PDO::FETCH_ASSOC);
 
-        $user_colors = [];
+            $user_colors = [];
 
-        if ($result_user) {
-            $query_user_colors = "SELECT * FROM user_colors WHERE user_id = :user_id";
-            $stmt_user_colors = $this->db_conn->prepare($query_user_colors);
-            $stmt_user_colors->bindParam(':user_id', $user_id, PDO::PARAM_INT);
-            $stmt_user_colors->execute();
+            if ($result_user) {
+                $query_user_colors = "SELECT * FROM user_colors WHERE user_id = :user_id";
+                $stmt_user_colors = $this->db_conn->prepare($query_user_colors);
+                $stmt_user_colors->bindParam(':user_id', $user_id, PDO::PARAM_INT);
+                $stmt_user_colors->execute();
 
-            $result_user_colors = $stmt_user_colors->fetchAll(PDO::FETCH_ASSOC);
+                $result_user_colors = $stmt_user_colors->fetchAll(PDO::FETCH_ASSOC);
 
-            if ($result_user_colors) {
-                foreach ($result_user_colors as $color_id) {
+                if ($result_user_colors) {
+                    foreach ($result_user_colors as $color_id) {
 
-                    $color = $colorDao->getColorById($color_id['color_id']);
-                    array_push($user_colors, $color);
+                        $color = $colorDao->getColorById($color_id['color_id']);
+                        array_push($user_colors, $color);
 
+                    }
                 }
-            }
 
-            return new User($result_user['name'], $result_user['email'], $result_user['id'], $user_colors);
+                return new User($result_user['name'], $result_user['email'], $result_user['id'], $user_colors);
+            } else {
+                throw new Exception("Usuário não encontrado.");
+            }
         } else {
-            throw new Exception("Usuário não encontrado.");
+            throw new Exception("Erro ao obter usuario no banco de dados.");
         }
 
     }
@@ -84,7 +87,7 @@ class UserDao
         $stmt_insert = $this->db_conn->prepare($insert_query);
         $stmt_insert->bindParam(':user_id', $user_id, PDO::PARAM_INT);
         $stmt_insert->bindParam(':color_id', $color_id, PDO::PARAM_INT);
-        $stmt_insert->execute();
+        return $stmt_insert->execute();
 
     }
 
@@ -93,7 +96,7 @@ class UserDao
         $delete_query = "DELETE FROM user_colors WHERE user_id = :user_id";
         $stmt_delete = $this->db_conn->prepare($delete_query);
         $stmt_delete->bindParam(':user_id', $user_id, PDO::PARAM_INT);
-        $stmt_delete->execute();
+        return $stmt_delete->execute();
     }
 
     public function update(User $user)
@@ -110,10 +113,14 @@ class UserDao
         $stmt_update->bindParam(':user_id', $user_id, PDO::PARAM_INT);
 
         if ($stmt_update->execute()) {
-            $this->deleteColorsForUserId($user_id);
-
-            foreach ($user->getColors() as $color) {
-                $this->insertColorForUserId($user_id, $color);
+            if ($this->deleteColorsForUserId($user_id)) {
+                foreach ($user->getColors() as $color) {
+                    if (!$this->insertColorForUserId($user_id, $color)) {
+                        throw new Exception("Erro ao atualizar usuário no banco de dados (insertColorForUserId). ");
+                    }
+                }
+            } else {
+                throw new Exception("Erro ao atualizar usuário no banco de dados (deleteColorsForUserId). ");
             }
         } else {
             throw new Exception("Erro ao atualizar usuário no banco de dados");
@@ -135,7 +142,9 @@ class UserDao
         if ($stmt_insert->execute()) {
             $user_id = $this->db_conn->lastInsertId();
             foreach ($user->getColors() as $color) {
-                $this->insertColorForUserId($user_id, $color);
+                if (!$this->insertColorForUserId($user_id, $color)) {
+                    throw new Exception("Erro ao inserir novo usuário no banco de dados (insertColorForUserId). ");
+                }
             }
         } else {
             throw new Exception("Erro ao inserir novo usuário no banco de dados");
@@ -147,13 +156,20 @@ class UserDao
     {
 
         $user_id = (int) $user->getId();
-        $this->deleteColorsForUserId($user_id);
 
-        $delete_query = "DELETE FROM users WHERE id = :user_id";
-        $stmt_delete = $this->db_conn->prepare($delete_query);
-        $stmt_delete->bindParam(':user_id', $user_id, PDO::PARAM_INT);
+        if ($this->deleteColorsForUserId($user_id)) {
+            $delete_query = "DELETE FROM users WHERE id = :user_id";
+            $stmt_delete = $this->db_conn->prepare($delete_query);
+            $stmt_delete->bindParam(':user_id', $user_id, PDO::PARAM_INT);
 
-        return $stmt_delete->execute();
+            if ($stmt_delete->execute()) {
+                return true;
+            } else {
+                throw new Exception("Erro ao deletar usuário do banco de dados.");
+            }
+        } else {
+            throw new Exception("Erro ao deletar usuário do banco de dados (deleteColorsForUserId) .");
+        }
 
     }
